requirements.txt
crispy-bootstrap5==2023.10
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
gunicorn==21.2.0
oauthlib==3.2.2
packaging==23.1
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.9.0
urllib3==1.26.14
whitenoise==6.6.0
pillow==10.4.0
django-silk==5.2.0
shortuuid>=1
hyperp>=1.0.4
email-validator==2.2.0
django_cors_headers==4.4.0
user_agents==2.2.0
markdown==3.7
